load-graph: Change variable from double to int

+++ src/load-graph.h
@@ -58,7 +58,7 @@ struct LoadGraph
     guint render_counter;
     guint frames_per_unit;
     guint graph_dely;
-    double graph_delx;
+    guint graph_delx;
     guint graph_buffer_offset;
     guint num_bars;
     guint real_draw_height;
